requirements.txt
jinja2
pyyaml
anthropic
